fix: NanoTDF secure key from debug logging and iv conflict risk
This change is motivated from the CodeQL result: https://github.com/opentdf/java-sdk/security/code-scanning/1

Although that use of a static IV is deliberate, it helped highlight that we should ensure that there is no reuse of the IV when encrypting the data.

In addition it was found that there were two places the key was logged, due to the sensitivity of the key this has been removed.