Fix test from minimum segment size enforcement
This change was applied easily in go, but there are issues with integration and other existing test payloads.
Because this is low risk, I believe it's ok to remove this protection in the Java SDK, but leave commented so it's known to be explict.  Alternatively we could update test payloads.