fix(sdk): upgrade the platform protocol code (#275)
In gateway we are seeing an error like
```
Caused by: com.google.protobuf.InvalidProtocolBufferException: While parsing a protocol message, the input ended unexpectedly in the middle of a field.  This could mean either that the input has been truncated or that an embedded message misreported its own length
```
when fetching attributes. [`v2.6.2` of platform contains `v0.5.0` of
`protocol/go`](https://github.com/virtru-corp/data-security-platform/blob/v2.6.2/go.mod#L34)
and there are changes between the two versions in the attributes. This
should at least be safe.