feat(sdk)!: add base key and support for key grants in ZTDF (#271)
Introduces the `Planner` class, which encapsulates the policy about how
to create a split plan, basing its
decision on `TDFConfig.autoconfigure`, whether KASes were provided in
config, and whether a split plan
was manually provided in config.

Modifies the `Autoconfigure.Granter` class to consume key grants. When a
key is specified using a KID
the KID is used during key assignment. When there are no grants or
mapped keys present `Autoconfigure.Granter`
reaches out to the `wellknown` endpoint to download a base key and
constructs a single element split plan.

* update visibility of some internal classes
* get rid of a bit of duplication
* include a `WellknownServiceClientInterface` in the `SDK.Services`
interface. Not sure if this is a good idea
or not but it doesn't seem dangerous.
* update some test classes